Revert "e2fsck: Change kmem_cache_t to lkmem_cache_t for Solaris"
This is a logical revert of commit 1911bf113ef0, for which the
description reads:

    Solaris polutes the C namespace with kmem_cache_t when
    you include in/netinet.h is included, so rename kmem_cache_t
    to lkmem_cache_t.

Reverting this change allows us to keep e2fsck/revoke.c in sync with
its upstream kernel source of fs/jbd2/revoke.c, and was the last
change required to make the e2fsprogs and kernel versions of revoke.c
to be bit identical.

Hopefully, a decade past from when this commit was originally
required, Solaris no longer needs this workaround for its buggy header
files leaking the structure definition of kmem_cache.  If it is still
needed there are other ways we can workaround this problem, assuming
there are still people in 2019 who care about building e2fsprogs for
Illumos / Open Solaris and its derivatives.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>